+++ src/server/modules/rest/impl/src/main/java/it/geosolutions/geostore/services/rest/utils/GeoStoreAuthenticationInterceptor.java
@@ -20,12 +20,8 @@
 package it.geosolutions.geostore.services.rest.utils;
 
 import it.geosolutions.geostore.core.model.User;
-import it.geosolutions.geostore.core.model.enums.Role;
 import it.geosolutions.geostore.services.UserService;
 import it.geosolutions.geostore.services.exception.NotFoundServiceEx;
-
-import java.util.List;
-import java.util.Map;
 
 import org.apache.cxf.interceptor.security.AccessDeniedException;
 import org.apache.cxf.message.Message;
@@ -33,39 +29,17 @@ import org.apache.cxf.message.Message;
 /**
  * 
  * Class GeoStoreAuthenticationInterceptor. Starting point was
- * JAASLoginInterceptor. Allow users auto creation
+ * JAASLoginInterceptor.
  * 
  * @author ETj (etj at geo-solutions.it)
  * @author Tobia di Pisa (tobia.dipisa at geo-solutions.it)
  * @author adiaz (alejandro.diaz at geo-solutions.it)
  */
 public class GeoStoreAuthenticationInterceptor extends
-		AbstractGeoStoreAuthenticationInterceptor {
-
+		AbstractGeoStoreAuthenticationInterceptor {	
+	
 	private UserService userService;
-
-	/**
-	 * Flag to indicate if an user that not exists could be created when it's
-	 * used
-	 */
-	private Boolean autoCreateUsers = false;
-
-	/**
-	 * Role for the new user
-	 */
-	private Role newUsersRole = Role.USER;
-
-	/**
-	 * New password strategy @see {@link NewPasswordStrategy}
-	 */
-	private NewPasswordStrategy newUsersPassword = NewPasswordStrategy.NONE;
-
-	/**
-	 * Header key for the new password if the selected strategy is
-	 * {@link NewPasswordStrategy#FROMHEADER}
-	 */
-	private String newUsersPasswordHeader = "";
-
+	
 	/**
 	 * @param userService
 	 *            the userService to set
@@ -73,50 +47,7 @@ public class GeoStoreAuthenticationInterceptor extends
 	public void setUserService(UserService userService) {
 		this.userService = userService;
 	}
-
-	public void setAutoCreateUsers(Boolean autoCreateUsers) {
-		this.autoCreateUsers = autoCreateUsers;
-	}
-
-	public void setNewUsersRole(Role newUsersRole) {
-		this.newUsersRole = newUsersRole;
-	}
-
-	public void setNewUsersPassword(NewPasswordStrategy newUsersPassword) {
-		this.newUsersPassword = newUsersPassword;
-	}
-
-	public void setNewUsersPasswordHeader(String newUsersPasswordHeader) {
-		this.newUsersPasswordHeader = newUsersPasswordHeader;
-	}
-
-	/**
-	 * Obtain the new password for a new user
-	 * 
-	 * @param message
-	 * @param username
-	 * 
-	 * @return password for the new user
-	 */
-	private String getNewUserPassword(Message message, String username) {
-		switch (newUsersPassword) {
-		case NONE:
-			return "";
-		case USERNAME:
-			return username;
-		case FROMHEADER:
-			@SuppressWarnings("unchecked")
-			Map<String, List<String>> headers = (Map<String, List<String>>) message
-					.get(Message.PROTOCOL_HEADERS);
-			if (headers.containsKey(newUsersPasswordHeader)) {
-				return headers.get(newUsersPasswordHeader).get(0);
-			}
-			return "";
-		default:
-			return "";
-		}
-	}
-
+	
 	/**
 	 * Obtain an user from his username
 	 * 
@@ -133,35 +64,15 @@ public class GeoStoreAuthenticationInterceptor extends
 			// Search on db
 			user = userService.get(username);
 		} catch (NotFoundServiceEx e) {
-			if (LOGGER.isInfoEnabled())
+			if (LOGGER.isInfoEnabled()){
 				LOGGER.info("Requested user not found: " + username);
-
-			// Auto create user
-			if (autoCreateUsers) {
-				if (LOGGER.isInfoEnabled()) {
-					LOGGER.info("Creating now");
-				}
-				user = new User();
-				user.setName(username);
-
-				user.setNewPassword(getNewUserPassword(message, username));
-				user.setRole(newUsersRole);
-				try {
-					// insert
-					user.setId(userService.insert(user));
-					// reload user stored
-					user = userService.get(username);
-				} catch (Exception e1) {
-					throw new AccessDeniedException(
-							"Not able to create new user");
-				}
-			} else {
-				throw new AccessDeniedException("Not authorized");
 			}
+			// throw exception
+			throw new AccessDeniedException("Not authorized");
 		}
-
+	
 		return user;
-
+	
 	}
 
 }
