+++ src/server/web/app/src/main/resources/geostore-ovr.properties
@@ -1,12 +1,19 @@
-# Uncomment the following to auto create unrecognized users
+#
+# Uncomment the following to auto create unrecognized users. 
+#
+# NOTE: You need to uncomment the interceptor on /src/modules/resp/impl/src/main/resources/applicationContext.xml:
+#         
+#	<jaxrs:inInterceptors>
+#      <!-- Auto create users interceptor (uncomment to allow users autocreation for /users requests) -->
+#      <ref bean="autoCreateUsersInterceptor"/>
 # 
-#geostoreAuthInterceptor.autoCreateUsers=true
+#autoCreateUsersInterceptor.autoCreateUsers=true
 #
 # Role to assign to automatically created users
 #
-#geostoreAuthInterceptor.newUsersRole.role=USER
+#autoCreateUsersInterceptor.newUsersRole.role=USER
 #
 # Strategy to assign password to automatically created users
 #
-#geostoreAuthInterceptor.newUsersPassword=NONE|USERNAME|FROMHEADER
-#geostoreAuthInterceptor.newUsersPasswordHeader=newUserPassword
+#autoCreateUsersInterceptor.newUsersPassword=NONE|USERNAME|FROMHEADER
+#autoCreateUsersInterceptor.newUsersPasswordHeader=newUserPassword
